Reafctor: extract RewriteSourceFileWrapper
